refactor: Exception 추가/수정 및 메서드 수정
- CustomException > PolicyException으로 변경
- DBException 추가
- ~Service: 각 서비스 조회에 Optional 추가
- ~Service: 테마, 예약시간이 예약에 사용될 경우 삭제 못하는 정책 적용
- ControllerAdvice에  변경된 Exception 클래스가 사용되도록 수정

+++ src/main/java/roomescape/admin/service/ReservationTimeService.java
@@ -5,8 +5,13 @@ import roomescape.admin.dto.ReadReservationTimeResponse;
 import roomescape.admin.dto.SaveReservationTimeRequest;
 import roomescape.admin.entity.ReservationTime;
 import roomescape.admin.repository.ReservationTimeRepository;
+import roomescape.common.exception.DBException;
+import roomescape.common.exception.PolicyException;
 
 import java.util.List;
+
+import static roomescape.common.exception.DBException.SEARCH_ERROR;
+import static roomescape.common.exception.PolicyException.RESERVATION_TIME_IN_USE_RESERVATION_ERROR;
 
 @Service
 public class ReservationTimeService {
@@ -18,19 +23,41 @@ public class ReservationTimeService {
     }
 
     public List<ReadReservationTimeResponse> readReservationTime() {
-        List<ReservationTime> reservationTime = this.reservationTimeRepository.findAll();
+        List<ReservationTime> reservationTime = this.reservationTimeRepository.findAll()
+                .orElseThrow(() -> new DBException(SEARCH_ERROR));
 
         return ReadReservationTimeResponse.from(reservationTime);
     }
 
     public ReadReservationTimeResponse saveReservationTime(SaveReservationTimeRequest saveReservationTimeRequest) {
         Long id = this.reservationTimeRepository.save(saveReservationTimeRequest);
-        ReservationTime reservationTime = this.reservationTimeRepository.findById(id);
 
-        return ReadReservationTimeResponse.from(reservationTime);
+        return ReadReservationTimeResponse.from(readReservationById(id));
+    }
+
+    public ReservationTime readReservationById(Long id) {
+        return this.reservationTimeRepository.findById(id)
+                .orElseThrow(() -> new DBException(SEARCH_ERROR));
     }
 
     public void deleteReservationTime(Long id) {
+        if(isContainReservation(id)){
+            throw new PolicyException(RESERVATION_TIME_IN_USE_RESERVATION_ERROR);
+        }
+
         this.reservationTimeRepository.delete(id);
+    }
+
+    private boolean isContainReservation(long reservationId) {
+        return isExistData(readReservationTimeCount(reservationId));
+    }
+
+    private Integer readReservationTimeCount(long reservationId) {
+        return this.reservationTimeRepository.countById(reservationId)
+                .orElseThrow(() -> new DBException(SEARCH_ERROR));
+    }
+
+    private boolean isExistData(int count){
+        return count > 0;
     }
 }
